+++ src/app/app-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from '@app/home/home.component';
+import { SelectivePreloadStrategyService } from '@app/route-utility/selective-preload-strategy.service';
 import { AuthenticationGuard } from '@auth/guards/authentication.guard';
 import { AppShellComponent } from '@layout/app-shell/app-shell.component';
 
@@ -9,7 +10,12 @@ const routes: Routes = [
     path: '',
     component: AppShellComponent,
     children: [
-      { path: 'home', component: HomeComponent },
+      {
+        path: 'home',
+        component: HomeComponent,
+        // TODO: Get proper page name & keys from SEO specialist and update pageTitle
+        data: { pageTitle: 'Home' }
+      },
       {
         path: 'auth',
         canLoad: [AuthenticationGuard],
@@ -20,6 +26,7 @@ const routes: Routes = [
       },
       {
         path: 'cart',
+        data: { preload: true, delay: true },
         loadChildren: () =>
           import('@app/cart/cart.module').then(m => m.CartModule)
       },
@@ -65,6 +72,8 @@ const routes: Routes = [
 @NgModule({
   imports: [
     RouterModule.forRoot(routes, {
+      enableTracing: true, // <-- debugging purposes only
+      preloadingStrategy: SelectivePreloadStrategyService,
       scrollPositionRestoration: 'enabled',
       anchorScrolling: 'enabled'
     })